Read file and flag missing values
The script opens the tab delimited data file in read format, it then reads the data portion of the file, row by row. The script searches for data rows which are missing the value for age. It flags those rows and informs the user of their length and diameter values.